xquery plugin log errors
