+++ plugins/projects/util/request_handler.py
@@ -16,5 +16,17 @@ class RequestHandler:
     def get_project(self):
         return self.session_payment.get('http://{}/create_project'.format(self.payment_processor_host), timeout=300).json()
 
+    def post_eth_proxy_project(self, host, data, project_id, api_key):
+        return self.session_eth.post('http://{}/xrs/evm_passthrough/{}'.format(host, project_id),
+                                     headers={
+                                         'Content-Type': 'application/json',
+                                         'Api-Key': api_key,
+                                     },
+                                     data=json.dumps(data), timeout=300).json()
+
     def list_projects(self):
         return self.session_payment.get('http://{}/list_projects'.format(self.payment_processor_host), timeout=300).json()
+
+    def post_update_api_count(self, project_id):
+        return self.session_payment.post('http://{}/{}/api_count'.format(self.payment_processor_host, project_id),
+                                         timeout=300)
